requirements.txt
flask
gunicorn
beautifulsoup4
readability-lxml
requests
deep-translator
lxml
